[product] : 주문과 RabbitMQ 연결

+++ product-service/src/main/java/com/oingmaryho/business/productservice/application/dto/response/ProductDetailsSearchResponseServiceDto.java
@@ -8,8 +8,8 @@ public record ProductDetailsSearchResponseServiceDto(
 	UUID companyId,
 	String companyName,
 	UUID manageHubId,
-	Long price,
-	Long stock
+	Integer price,
+	Integer stock
 
 ) {
 }
